requirements.txt
streamlit>=1.37.0
docx-mailmerge>=0.1.0
pikepdf>=8.0.0
pandas>=2.0.0
openpyxl>=3.1.0
